fix : subscription DTO file, role
