+++ start/src/main/java/com/effitizer/start/domain/dto/Subscribe/SubscribeDTO.java
@@ -9,15 +9,27 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor
 public class SubscribeDTO {
-    private Long id;
-    private LocalDateTime start_date;
-    private LocalDateTime expired_date;
-    private LocalDateTime canceled_date;
+    private Long id; //  id
 
-    public SubscribeDTO(Subscribe subscribe){
-        this.id = subscribe.getId();
-        this.start_date = subscribe.getStart_date();
-        this.expired_date = subscribe.getExpired_date();
-        this.canceled_date = subscribe.getCanceled_data();
+    private Long user_id; // 회원 id
+
+    private LocalDateTime start_date; //구독시작일
+    private LocalDateTime expired_date; //만료일
+    private LocalDateTime canceled_data; // 결제 중단일
+
+    public SubscribeDTO(Long id, Long user_id, LocalDateTime start_date, LocalDateTime expired_date, LocalDateTime canceled_data){
+        this.id=id;
+        this.user_id=user_id;
+        this.start_date=start_date;
+        this.expired_date=expired_date;
+        this.canceled_data=canceled_data;
+    }
+
+    public SubscribeDTO(Subscribe subscribe) {
+        this.id=subscribe.getId();
+        this.user_id=subscribe.getUser().getId();
+        this.start_date=subscribe.getStart_date();
+        this.expired_date=subscribe.getExpired_date();
+        this.canceled_data=subscribe.getCanceled_data();
     }
 }
