+++ start/src/main/java/com/effitizer/start/service/SubscribeService.java
@@ -28,21 +28,24 @@ public class SubscribeService {
     @Autowired
     SubscribeRepository subscribeRepository;
 
+    //user 아이디로 구독 정보 기록 조회
     public List<Subscribe> findSubscribesByUserId(Long user_id) {
         return subscribeRepository.findByUserId(user_id);
     }
 
+    //user 아이디로 구독 정보 조회
     public Subscribe findSubscribeByUserId(Long user_id) {
         return subscribeRepository.findFirstByUserIdOrderByIdDesc(user_id)
                 .orElse(null);
     }
+    //user 구독 추가
     public void saveOne(User user) {
 
         LocalDateTime current = LocalDateTime.now();
         Subscribe subscribe =new Subscribe(user,current,current.plusMonths(1),current.plusMonths(1));
         subscribeRepository.save(subscribe);
     }
-
+    //user 구독 연장
     public void updateOne(User user, LocalDateTime expired_date) {
 
         LocalDateTime current = LocalDateTime.now();
